Refactored Converters from Views
